+++ ChangeLog
@@ -1,0 +1,4 @@
+
+| Date | Description | Author | Username |
+| ---- | ----------- | ------ | -------- |
+| 04-Jun-2019 | Refactored project and build system | Jose Fernando Lopez Fernandez | [jflopezfernandez](https://github.com/) |
