--- Makefile.am
@@ -1,4 +1,0 @@
-
-SUBDIRS=src
-man1_MANS=dcp.man
-EXTRA_DIST=dcp.man
