+++ src/main.c
@@ -31,11 +31,11 @@
 #include "config.h"     /* generated by autotools */
 #include "cmdline.h"    /* generated by gengetopt */
 #include "digest.h"
-#include "index/index.h"
-#include "io/io.h"
+#include "index.h"
+#include "io.h"
 #include "io_dcp_processor.h"
 #include "logging.h"
-#include "impl/dcp.h"
+#include "dcp.h"
 
 
 /* MACROS *********************************************************************/
